[#31] feat: order response 생성하는 로직 이동, 서비스단에서 컨트롤러로
