+++ src/main/java/com/itda/apiserver/domain/ShippingInfo.java
@@ -14,7 +14,7 @@ public class ShippingInfo extends Core {
     @Embedded
     private Address address;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
